Cap the endline when formatting docstrings.
`Newline` nodes typically by checking whether the startline of the
following node is greater than the endline of the previous node. When a
docstring is formatted it can change the number of the lines in the
docstring which is problematic since the formatter expects the number of
lines shown in the source file.